Fix underscores in link being formatted as markdown
In the existing version, the link itself works but is cut off after `should`, and then `stackage` and `version` are italicized, because of the underscores. I'm not sure if this is the best solution—something is clearly finding text that looks like links and making them links, and maybe the better solution is to make that not generate weird link text.